Initial deserialization call plus refactoring
The message derserializer creates the top level payload object and
passes it off to the helper method that walks the JSON in conjunction
with the property lists from the serializable payload type. Also
refactored the package paths to make the de-coupling clearer regarding
the visibility of classes in the compilation and runtime modules.
