+++ generic-messages/src/main/java/com/langtoun/messages/compilation/properties/ScalarProperty.java
@@ -1,4 +1,4 @@
-package com.langtoun.messages.properties;
+package com.langtoun.messages.compilation.properties;
 
 import java.util.function.Consumer;
 import java.util.function.Supplier;
