+++ generic-messages/src/main/java/com/langtoun/messages/compilation/types/cars/ComplexCar.java
@@ -1,10 +1,10 @@
-package com.langtoun.messages.types.cars;
+package com.langtoun.messages.compilation.types.cars;
 
-import static com.langtoun.messages.properties.ScalarProperty.newScalarProperty;
+import static com.langtoun.messages.compilation.properties.ScalarProperty.newScalarProperty;
 
 import java.util.List;
 
-import com.langtoun.messages.properties.MessageProperty;
+import com.langtoun.messages.compilation.properties.MessageProperty;
 
 public class ComplexCar extends SimpleCar {
